// api/generated/models/createUsuarioDtoTipo.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Unibac API
 * Documentación de la API
 * OpenAPI spec version: 1.0
 */

export type CreateUsuarioDtoTipo =
  (typeof CreateUsuarioDtoTipo)[keyof typeof CreateUsuarioDtoTipo];

export const CreateUsuarioDtoTipo = {
  INTERNO: "INTERNO",
  EXTERNO: "EXTERNO",
} as const;